+++ prover/src/utils.rs
@@ -2,7 +2,8 @@
 //! operations.
 
 use binius_field::ExtensionField;
-use binius_m3::builder::{upcast_expr, Col, Expr, TableBuilder, B128, B16, B32};
+use binius_field::Field;
+use binius_m3::builder::{upcast_col, upcast_expr, Col, Expr, TableBuilder, B1, B128, B16, B32};
 
 /// Get a B128 basis element by index
 #[inline]
@@ -167,4 +168,38 @@ pub fn pack_instruction_with_32bits_imm_b128(pc: B32, opcode: B16, arg: B16, imm
 /// Packs two 16-bit limbs into a single 32-bit value.
 pub(crate) fn pack_b16_into_b32(low: Col<B16, 1>, high: Col<B16, 1>) -> Expr<B32, 1> {
     upcast_expr(high.into()) * <B32 as ExtensionField<B16>>::basis(1) + upcast_expr(low.into())
+}
+
+/// Sets up a multiplexer constraint for bit selection.
+///
+/// Creates a constraint that enforces: result = select_bit ? when_true :
+/// when_false
+///
+/// # Arguments
+///
+/// * `table` - The table builder to add the constraint to
+/// * `result` - The result column that will store the selected value
+/// * `when_true` - The column to select when select_bit is 1
+/// * `when_false` - The column to select when select_bit is 0
+/// * `select_bit` - The control bit that determines which value to select
+pub fn setup_mux_constraint<const N: usize>(
+    table: &mut TableBuilder,
+    result: &Col<B1, N>,
+    when_true: &Col<B1, N>,
+    when_false: &Col<B1, N>,
+    select_bit: &Col<B1>,
+) {
+    // Create packed versions of columns
+    let result_packed = table.add_packed("result_packed", *result);
+    let true_packed = table.add_packed("when_true_packed", *when_true);
+    let false_packed = table.add_packed("when_false_packed", *when_false);
+
+    // Create constraint for the mux:
+    // result = select_bit ? when_true : when_false
+    table.assert_zero(
+        "mux_constraint",
+        result_packed
+            - (true_packed * upcast_col(*select_bit)
+                + false_packed * (upcast_col(*select_bit) - B32::ONE)),
+    );
 }
